fix: 경매 서비스 코드 분리 및 예외 구체화
